requirements.txt
# Core data processing
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
requests==2.31.0

# AWS
boto3==1.34.0

# Database
psycopg2-binary==2.9.9

# Dashboard
streamlit==1.31.0

# Visualization (used in Colab notebook)
matplotlib==3.8.2
